ShortCode implementations moved into main exe
