+++ BookGen.Tests/UT_ShortCodeLoader.cs
@@ -1,10 +1,10 @@
 ﻿//-----------------------------------------------------------------------------
-// (c) 2019 Ruzsinszki Gábor
+// (c) 2019-2020 Ruzsinszki Gábor
 // This code is licensed under MIT license (see LICENSE for details)
 //-----------------------------------------------------------------------------
 
 using BookGen.Core;
-using BookGen.Template.ShortCodeImplementations;
+using BookGen.Framework.Shortcodes;
 using BookGen.Tests.Environment;
 using NUnit.Framework;
 using System.Linq;
