+++ BookGen.Tests/UT_Translate.cs
@@ -1,12 +1,12 @@
 ﻿//-----------------------------------------------------------------------------
-// (c) 2019 Ruzsinszki Gábor
+// (c) 2019-2020 Ruzsinszki Gábor
 // This code is licensed under MIT license (see LICENSE for details)
 //-----------------------------------------------------------------------------
 
-using BookGen.Template.ShortCodeImplementations;
 using BookGen.Core.Configuration;
 using NUnit.Framework;
 using System.Collections.Generic;
+using BookGen.Framework.Shortcodes;
 
 namespace BookGen.Tests
 {
