+++ BookGen/Framework/Shortcodes/BuildTime.cs
@@ -1,5 +1,5 @@
 ﻿//-----------------------------------------------------------------------------
-// (c) 2019 Ruzsinszki Gábor
+// (c) 2019-2020 Ruzsinszki Gábor
 // This code is licensed under MIT license (see LICENSE for details)
 //-----------------------------------------------------------------------------
 
@@ -8,7 +8,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 
-namespace BookGen.Template.ShortCodeImplementations
+namespace BookGen.Framework.Shortcodes
 {
     [Export(typeof(ITemplateShortCode))]
     public class BuildTime : ITemplateShortCode
