+++ BookGen/Framework/Shortcodes/InlineFile.cs
@@ -1,5 +1,5 @@
 ﻿//-----------------------------------------------------------------------------
-// (c) 2019 Ruzsinszki Gábor
+// (c) 2019-2020 Ruzsinszki Gábor
 // This code is licensed under MIT license (see LICENSE for details)
 //-----------------------------------------------------------------------------
 
@@ -9,7 +9,7 @@ using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using BookGen.Core.Contracts;
 
-namespace BookGen.Template.ShortCodeImplementations
+namespace BookGen.Framework.Shortcodes
 {
     [Export(typeof(ITemplateShortCode))]
     public class InlineFile : ITemplateShortCode
