+++ BookGen/Framework/Shortcodes/JSPageToc.cs
@@ -1,14 +1,15 @@
 ﻿//-----------------------------------------------------------------------------
-// (c) 2019 Ruzsinszki Gábor
+// (c) 2019-2020 Ruzsinszki Gábor
 // This code is licensed under MIT license (see LICENSE for details)
 //-----------------------------------------------------------------------------
 
 using BookGen.Core;
 using BookGen.Core.Contracts;
+using BookGen.Template;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 
-namespace BookGen.Template.ShortCodeImplementations
+namespace BookGen.Framework.Shortcodes
 {
     [Export(typeof(ITemplateShortCode))]
     public class JSPageToc : ITemplateShortCode
