+++ Bookgen.Template/Bookgen.Template.csproj
@@ -63,19 +63,15 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="System.ComponentModel.Composition" Version="4.7.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\BookGen.Core\BookGen.Core.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
     <Folder Include="Properties\" />
   </ItemGroup>
 
   <ItemGroup>
     <EmbeddedResource Include="TemplateScript\ScriptTemplate.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BookGen.Core\BookGen.Core.csproj" />
   </ItemGroup>
 
 </Project>
